+++ tests/command-line-options.test.ts
@@ -421,27 +421,6 @@ test.only("external is string?", () => {
 
       expect(status).toBe(0);
     });
-
-    test("handles not supported '--target' option value", async () => {
-      await writeFixture(fixture, {
-        ["__typetests__/dummy.test.ts"]: isStringTestText,
-        ["tsconfig.json"]: JSON.stringify(tsconfig, null, 2),
-      });
-
-      const { status, stderr, stdout } = spawnTyche(fixture, ["--target", "new"]);
-
-      expect(stdout).toBe("");
-      expect(stderr).toMatch(
-        [
-          "Error: TypeScript version 'new' is not supported.",
-          "",
-          "Argument for the '--target' option must be a single tag or a comma separated list.",
-          "Usage examples:",
-        ].join("\r\n"),
-      );
-
-      expect(status).toBe(1);
-    });
   });
 
   describe("'--update' option", () => {
